+++ controllers/keda/scaledjob_controller.go
@@ -191,6 +191,11 @@ func (r *ScaledJobReconciler) reconcileScaledJob(ctx context.Context, logger log
 		return "ScaledJob doesn't have correct triggers specification", err
 	}
 
+	err = r.updateStatusWithTriggersAndAuthsNames(ctx, logger, scaledJob)
+	if err != nil {
+		return "Cannot update ScaledJob status with triggers'names and authentications'names", err
+	}
+
 	// nosemgrep: trailofbits.go.invalid-usage-of-modified-variable.invalid-usage-of-modified-variable
 	msg, err := r.deletePreviousVersionScaleJobs(ctx, logger, scaledJob)
 	if err != nil {
@@ -403,4 +408,14 @@ func (r *ScaledJobReconciler) updateTriggerAuthenticationStatusOnDelete(ctx cont
 		triggerAuthenticationStatus.ScaledJobNamesStr = kedacontrollerutil.RemoveFromString(triggerAuthenticationStatus.ScaledJobNamesStr, scaledJob.GetName(), ",")
 		return triggerAuthenticationStatus
 	})
+}
+
+func (r *ScaledJobReconciler) updateStatusWithTriggersAndAuthsNames(ctx context.Context, logger logr.Logger, scaledJob *kedav1alpha1.ScaledJob) error {
+	triggersNames, authsNames := kedav1alpha1.GetCombinedTriggersAndAuthenticationsNames(scaledJob.Spec.Triggers)
+	status := scaledJob.Status.DeepCopy()
+	status.TriggersNames = &triggersNames
+	status.AuthenticationsNames = &authsNames
+
+	logger.Info("Updating ScaledJob status with triggers and authentications names", "triggersNames", triggersNames, "authenticationsNames", authsNames)
+	return kedastatus.UpdateScaledJobStatus(ctx, r.Client, logger, scaledJob, status)
 }