+++ controllers/keda/scaledobject_controller.go
@@ -282,6 +282,11 @@ func (r *ScaledObjectReconciler) reconcileScaledObject(ctx context.Context, logg
 		return "ScaledObject doesn't have correct triggers specification", err
 	}
 
+	err = r.updateStatusWithTriggersAndAuthsNames(ctx, logger, scaledObject)
+	if err != nil {
+		return "Cannot update ScaledObject status with triggers'names and authentications'names", err
+	}
+
 	// Create a new HPA or update existing one according to ScaledObject
 	newHPACreated, err := r.ensureHPAForScaledObjectExists(ctx, logger, scaledObject, &gvkr)
 	if err != nil {
@@ -620,4 +625,12 @@ func (r *ScaledObjectReconciler) updateTriggerAuthenticationStatusOnDelete(ctx c
 			triggerAuthenticationStatus.ScaledObjectNamesStr = kedacontrollerutil.RemoveFromString(triggerAuthenticationStatus.ScaledObjectNamesStr, scaledObject.GetName(), ",")
 			return triggerAuthenticationStatus
 		})
+}
+
+func (r *ScaledObjectReconciler) updateStatusWithTriggersAndAuthsNames(ctx context.Context, logger logr.Logger, scaledObject *kedav1alpha1.ScaledObject) error {
+	triggersNames, authsNames := kedav1alpha1.GetCombinedTriggersAndAuthenticationsNames(scaledObject.Spec.Triggers)
+	status := scaledObject.Status.DeepCopy()
+	status.TriggersNames = &triggersNames
+	status.AuthenticationsNames = &authsNames
+	return kedastatus.UpdateScaledObjectStatus(ctx, r.Client, logger, scaledObject, status)
 }